roachtest: move PredecessorVersions to `tests`
Avoids import conflict as more roachtests move to `tests`.

Release note: None
